requirements.txt
mysql-connector-python==8.0.20
pytest==5.4.3
pytest-xdist==1.32.0
requests==2.24.0
selenium==3.141.0